refactor: merging the ‘setupTest’ helpers into LearningHeader.test.jsx
setupTest was just exporting stuff being used by the LearningHeader.test.jsx file.  No other files were using it, and it wasn’t actually a ‘setupTest’ file so much as a library of helper functions.  This just moves them all into the file and deletes setupTest.
